+++ src/components/form-title/form-title.validations.test.js
@@ -1,5 +1,5 @@
 import { schema } from './form-title.validations';
-import datasets from '../../../../test/fixtures/datasets';
+import datasets from '../../../test/fixtures/datasets';
 
 it('should validate without errors', async () => {
   const dataset = datasets.datasetItems._embedded.datasets[0];
